+++ src/Entities/UPA/index.ts
@@ -2,5 +2,6 @@ export * from "./ExtraChargeType";
 export * from "./Lodging";
 export * from "./POSData";
 export * from "./ProcessIndicator";
+export * from "./ScanData";
 export * from "./UpaParam";
 export * from "./UpaTransactionData";